requirements.txt
Django==6.0.2
djangorestframework==3.16.1
python-dotenv==1.2.1
requests==2.32.5
